requirement.txt
streamlit-calendar
ics
streamlit
jq
pytz
python-dotenv
requests
faiss-cpu
openai
langchain
langchain-text-splitters